+++ zoomimage-compose/src/commonTest/kotlin/com/github/panpf/zoomimage/compose/common/test/subsampling/SubsamplingStateTest.kt
@@ -6,6 +6,7 @@ import androidx.compose.runtime.snapshotFlow
 import androidx.compose.ui.test.ExperimentalTestApi
 import androidx.compose.ui.test.runComposeUiTest
 import androidx.compose.ui.unit.IntSize
+import androidx.compose.ui.unit.LayoutDirection
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import com.githb.panpf.zoomimage.images.ResourceImages
 import com.github.panpf.zoomimage.compose.rememberZoomImageLogger
@@ -53,7 +54,7 @@ class SubsamplingStateTest {
     @Test
     fun testConstructor() {
         val logger = Logger("Test")
-        val zoomable = ZoomableState(logger)
+        val zoomable = ZoomableState(logger, LayoutDirection.Ltr)
         val testLifecycle = TestLifecycle()
         val subsampling = SubsamplingState(zoomable, testLifecycle)
         assertSame(expected = zoomable, actual = subsampling.zoomableState)
@@ -65,7 +66,7 @@ class SubsamplingStateTest {
     @Test
     fun testLogger() {
         val logger = Logger("Test")
-        val zoomable = ZoomableState(logger)
+        val zoomable = ZoomableState(logger, LayoutDirection.Ltr)
         val testLifecycle = TestLifecycle()
         val subsampling = SubsamplingState(zoomable, testLifecycle)
         assertSame(expected = logger, actual = subsampling.logger)
@@ -74,7 +75,7 @@ class SubsamplingStateTest {
     @Test
     fun testTileImageCache() {
         val logger = Logger("Test")
-        val zoomable = ZoomableState(logger)
+        val zoomable = ZoomableState(logger, LayoutDirection.Ltr)
         val testLifecycle = TestLifecycle()
         val subsampling = SubsamplingState(zoomable, testLifecycle)
         assertEquals(expected = null, actual = subsampling.tileImageCache)
@@ -87,7 +88,7 @@ class SubsamplingStateTest {
     @Test
     fun testDisabledTileImageCache() {
         val logger = Logger("Test")
-        val zoomable = ZoomableState(logger)
+        val zoomable = ZoomableState(logger, LayoutDirection.Ltr)
         val testLifecycle = TestLifecycle()
         val subsampling = SubsamplingState(zoomable, testLifecycle)
         assertEquals(expected = false, actual = subsampling.disabledTileImageCache)
@@ -99,7 +100,7 @@ class SubsamplingStateTest {
     @Test
     fun testTileAnimationSpec() {
         val logger = Logger("Test")
-        val zoomable = ZoomableState(logger)
+        val zoomable = ZoomableState(logger, LayoutDirection.Ltr)
         val testLifecycle = TestLifecycle()
         val subsampling = SubsamplingState(zoomable, testLifecycle)
         assertEquals(expected = TileAnimationSpec.Default, actual = subsampling.tileAnimationSpec)
@@ -111,7 +112,7 @@ class SubsamplingStateTest {
     @Test
     fun testPausedContinuousTransformTypes() {
         val logger = Logger("Test")
-        val zoomable = ZoomableState(logger)
+        val zoomable = ZoomableState(logger, LayoutDirection.Ltr)
         val testLifecycle = TestLifecycle()
         val subsampling = SubsamplingState(zoomable, testLifecycle)
         assertEquals(
@@ -130,7 +131,7 @@ class SubsamplingStateTest {
     @Test
     fun testDisabledBackgroundTiles() {
         val logger = Logger("Test")
-        val zoomable = ZoomableState(logger)
+        val zoomable = ZoomableState(logger, LayoutDirection.Ltr)
         val testLifecycle = TestLifecycle()
         val subsampling = SubsamplingState(zoomable, testLifecycle)
         assertEquals(expected = false, actual = subsampling.disabledBackgroundTiles)
@@ -142,7 +143,7 @@ class SubsamplingStateTest {
     @Test
     fun testStopped() {
         val logger = Logger("Test")
-        val zoomable = ZoomableState(logger)
+        val zoomable = ZoomableState(logger, LayoutDirection.Ltr)
         val testLifecycle = TestLifecycle()
         val subsampling = SubsamplingState(zoomable, testLifecycle)
         assertEquals(expected = false, actual = subsampling.stopped)
@@ -154,7 +155,7 @@ class SubsamplingStateTest {
     @Test
     fun testShowTileBounds() {
         val logger = Logger("Test")
-        val zoomable = ZoomableState(logger)
+        val zoomable = ZoomableState(logger, LayoutDirection.Ltr)
         val testLifecycle = TestLifecycle()
         val subsampling = SubsamplingState(zoomable, testLifecycle)
         assertEquals(expected = false, actual = subsampling.showTileBounds)