+++ zoomimage-compose/src/commonTest/kotlin/com/github/panpf/zoomimage/compose/common/test/zoom/ZoomableStateTest.kt
@@ -11,6 +11,7 @@ import androidx.compose.ui.test.runComposeUiTest
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.IntRect
 import androidx.compose.ui.unit.IntSize
+import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.toIntRect
 import androidx.compose.ui.unit.toOffset
 import androidx.compose.ui.unit.toSize
@@ -58,7 +59,7 @@ class ZoomableStateTest {
     @Test
     fun testConstructor() {
         val logger = Logger("Test")
-        val zoomable = ZoomableState(logger)
+        val zoomable = ZoomableState(logger, LayoutDirection.Ltr)
         assertEquals(expected = logger, actual = zoomable.logger)
         @Suppress("USELESS_IS_CHECK")
         assertEquals(expected = true, actual = zoomable is RememberObserver)
@@ -66,7 +67,7 @@ class ZoomableStateTest {
 
     @Test
     fun testContainerSize() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = IntSize.Zero, actual = zoomable.containerSize)
 
         zoomable.containerSize = IntSize(1000, 2000)
@@ -75,7 +76,7 @@ class ZoomableStateTest {
 
     @Test
     fun testContentSize() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = IntSize.Zero, actual = zoomable.contentSize)
 
         zoomable.containerSize = IntSize(1000, 2000)
@@ -89,7 +90,7 @@ class ZoomableStateTest {
 
     @Test
     fun testContentOriginSize() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = IntSize.Zero, actual = zoomable.contentOriginSize)
 
         zoomable.contentOriginSize = IntSize(4000, 2400)
@@ -98,7 +99,7 @@ class ZoomableStateTest {
 
     @Test
     fun testContentScale() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = ContentScale.Fit, actual = zoomable.contentScale)
 
         zoomable.contentScale = ContentScale.Crop
@@ -107,16 +108,18 @@ class ZoomableStateTest {
 
     @Test
     fun testAlignment() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = Alignment.Center, actual = zoomable.alignment)
 
         zoomable.alignment = Alignment.TopStart
         assertEquals(expected = Alignment.TopStart, actual = zoomable.alignment)
+
+        // TODO test LayoutDirection
     }
 
     @Test
     fun testReadMode() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = null, actual = zoomable.readMode)
 
         zoomable.readMode = ReadMode.Default
@@ -125,7 +128,7 @@ class ZoomableStateTest {
 
     @Test
     fun testScalesCalculator() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = ScalesCalculator.Dynamic, actual = zoomable.scalesCalculator)
 
         zoomable.scalesCalculator = ScalesCalculator.Fixed
@@ -134,7 +137,7 @@ class ZoomableStateTest {
 
     @Test
     fun testThreeStepScale() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = false, actual = zoomable.threeStepScale)
 
         zoomable.threeStepScale = true
@@ -143,7 +146,7 @@ class ZoomableStateTest {
 
     @Test
     fun testRubberBandScale() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = true, actual = zoomable.rubberBandScale)
 
         zoomable.rubberBandScale = false
@@ -152,7 +155,7 @@ class ZoomableStateTest {
 
     @Test
     fun testOneFingerScaleSpec() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = OneFingerScaleSpec.Default, actual = zoomable.oneFingerScaleSpec)
 
         zoomable.oneFingerScaleSpec = OneFingerScaleSpec(DefaultPanToScaleTransformer(100))
@@ -164,7 +167,7 @@ class ZoomableStateTest {
 
     @Test
     fun testAnimationSpec() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = ZoomAnimationSpec.Default, actual = zoomable.animationSpec)
 
         zoomable.animationSpec = ZoomAnimationSpec(durationMillis = 4000)
@@ -176,7 +179,7 @@ class ZoomableStateTest {
 
     @Test
     fun testLimitOffsetWithinBaseVisibleRect() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = false, actual = zoomable.limitOffsetWithinBaseVisibleRect)
 
         zoomable.limitOffsetWithinBaseVisibleRect = true
@@ -184,8 +187,14 @@ class ZoomableStateTest {
     }
 
     @Test
+    fun testContainerWhitespace() {
+        // TODO test
+        // TODO test LayoutDirection
+    }
+
+    @Test
     fun testDisabledGestureTypes() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = 0, actual = zoomable.disabledGestureTypes)
 
         zoomable.disabledGestureTypes = GestureType.ONE_FINGER_SCALE or GestureType.TWO_FINGER_SCALE
@@ -197,7 +206,7 @@ class ZoomableStateTest {
 
     @Test
     fun testReverseMouseWheelScale() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(expected = false, actual = zoomable.reverseMouseWheelScale)
 
         zoomable.reverseMouseWheelScale = true
@@ -206,7 +215,7 @@ class ZoomableStateTest {
 
     @Test
     fun testMouseWheelScaleScrollDeltaConverter() {
-        val zoomable = ZoomableState(Logger("Test"))
+        val zoomable = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(
             expected = 2.31f,
             actual = zoomable.mouseWheelScaleScrollDeltaConverter(7f).format(2)
@@ -3711,7 +3720,7 @@ class ZoomableStateTest {
 
     @Test
     fun testCheckSupportGestureType() {
-        val zoomableState = ZoomableState(Logger("Test"))
+        val zoomableState = ZoomableState(Logger("Test"), LayoutDirection.Ltr)
         assertEquals(
             expected = "[ONE_FINGER_DRAG:true, TWO_FINGER_SCALE:true, ONE_FINGER_SCALE:true, DOUBLE_TAP_SCALE:true, MOUSE_WHEEL_SCALE:true, KEYBOARD_SCALE:true, KEYBOARD_DRAG:true]",
             actual = GestureType.values.joinToString(prefix = "[", postfix = "]") {